+++ packages/sdk/src/Portfolio.ts
@@ -12,8 +12,8 @@ import { Viem } from "./Utils.js";
 export * as ExternalPosition from "./Portfolio/ExternalPosition.js";
 export * as ExternalPositions from "./Portfolio/ExternalPositions.js";
 
-export * as IntegrationAdapter from "./Portfolio/IntegrationAdapter.js";
-export * as IntegrationAdapters from "./Portfolio/IntegrationAdapters.js";
+export * as Integration from "./Portfolio/Integration.js";
+export * as Integrations from "./Portfolio/Integrations.js";
 
 export * as VoteDelegation from "./Portfolio/VoteDelegation.js";
