--- packages/sdk/src/Portfolio/IntegrationAdapters.ts
@@ -1,19 +1,0 @@
-export * as AaveV2 from "./IntegrationAdapters/AaveV2.js";
-export * as AaveV3 from "./IntegrationAdapters/AaveV3.js";
-export * as ArrakisV2 from "./IntegrationAdapters/ArrakisV2.js";
-export * as BalancerV2Liquidity from "./IntegrationAdapters/BalancerV2Liquidity.js";
-export * as CompoundV2 from "./IntegrationAdapters/CompoundV2.js";
-export * as CompoundV3 from "./IntegrationAdapters/CompoundV3.js";
-export * as ConvexCurveLpStaking from "./IntegrationAdapters/ConvexCurveLpStaking.js";
-export * as CurveExchange from "./IntegrationAdapters/CurveExchange.js";
-export * as CurveLiquidity from "./IntegrationAdapters/CurveLiquidity.js";
-export * as Erc4626 from "./IntegrationAdapters/Erc4626.js";
-export * as IdleV4 from "./IntegrationAdapters/IdleV4.js";
-export * as OneInchV5 from "./IntegrationAdapters/OneInchV5.js";
-export * as ParaswapV5 from "./IntegrationAdapters/ParaswapV5.js";
-export * as UniswapV2Exchange from "./IntegrationAdapters/UniswapV2Exchange.js";
-export * as UniswapV2Liquidity from "./IntegrationAdapters/UniswapV2Liquidity.js";
-export * as UniswapV3 from "./IntegrationAdapters/UniswapV3.js";
-export * as YearnVaultV2 from "./IntegrationAdapters/YearnVaultV2.js";
-export * as ZeroExV2 from "./IntegrationAdapters/ZeroExV2.js";
-export * as ZeroExV4 from "./IntegrationAdapters/ZeroExV4.js";
